requirements.txt
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.34
alembic==1.13.1
pytest==8.0.0
pydantic==2.6.3
fastapi==0.110.0
uvicorn==0.27.1 